Consuming API with React

+++ BONUS_RestWithASPNETUdemy_React/client/src/pages/Books/index.js
@@ -11,31 +11,32 @@ import logoImage from '../../assets/logo.svg'
 export default function Books(){
 
     const [books, setBooks] = useState([]);
+    const [page, setPage] = useState(1);
 
     const userName = localStorage.getItem('userName');
-
     const accessToken = localStorage.getItem('accessToken');
+
+    const authorization = {
+        headers: {
+            Authorization: `Bearer ${accessToken}`
+        }
+    }
 
     const history = useHistory();
 
     useEffect(() => {
-        api.get('api/Book/v1/asc/20/1', {
-            headers: {
-                Authorization: `Bearer ${accessToken}`
-            }
-        }).then(response => {
-            setBooks(response.data.list)
-        })
+        fetchMoreBooks();
     }, [accessToken]);
 
+    async function fetchMoreBooks() {
+        const response = await api.get(`api/Book/v1/asc/4/${page}`, authorization);
+        setBooks([ ...books, ...response.data.list]);
+        setPage(page + 1);
+    }
     
     async function logout() {
         try {
-            await api.get('api/auth/v1/revoke', {
-                headers: {
-                    Authorization: `Bearer ${accessToken}`
-                }
-            });
+            await api.get('api/auth/v1/revoke', authorization);
 
             localStorage.clear();
             history.push('/');
@@ -43,14 +44,18 @@ export default function Books(){
             alert('Logout failed! Try again!');
         }
     }
+    
+    async function editBook(id) {
+        try {
+            history.push(`book/new/${id}`)
+        } catch (err) {
+            alert('Edit book failed! Try again!');
+        }
+    }
 
     async function deleteBook(id) {
         try {
-            await api.delete(`api/Book/v1/${id}`, {
-                headers: {
-                    Authorization: `Bearer ${accessToken}`
-                }
-            });
+            await api.delete(`api/Book/v1/${id}`, authorization);
 
             setBooks(books.filter(book => book.id !== id))
         } catch (err) {
@@ -63,7 +68,7 @@ export default function Books(){
             <header>
                 <img src={logoImage} alt="Erudio"/>
                 <span>Welcome, <strong>{userName.toLowerCase()}</strong>!</span>
-                <Link className="button" to="book/new">Add New Book</Link>
+                <Link className="button" to="book/new/0">Add New Book</Link>
                 <button onClick={logout} type="button">
                     <FiPower size={18} color="#251FC5" />
                 </button>
@@ -82,7 +87,7 @@ export default function Books(){
                         <strong>Release Date:</strong>
                         <p>{Intl.DateTimeFormat('pt-BR').format(new Date(book.launchDate))}</p>
                         
-                        <button type="button">
+                        <button onClick={() => editBook(book.id)} type="button">
                             <FiEdit size={20} color="#251FC5"/>
                         </button>
                         
@@ -92,6 +97,7 @@ export default function Books(){
                     </li>
                 ))}
             </ul>
+            <button className="button" onClick={fetchMoreBooks} type="button">Load More</button>
         </div>
     );
 }
